TagDetailController moved to new folder Tag 	renamed:    app/Http/Controllers/api/TagDetailController.php -> app/Http/Controllers/api/Tag/TagDetailController.php 	modified:   routes/api/v1.php 	modified:   tests/Feature/TagDetailControllerTest.php
